% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{adBackEnd}
\alias{.adBackEnd}
\alias{adBackEnd}
\title{The adTestStatistic with ad.test as a potential backend}
\usage{
.adBackEnd(adjusted.y, z)
}
\arguments{
\item{adjusted.y}{y}

\item{z}{z}
}
\value{
RandomizationDistribution
}
\description{
The adTestStatistic with ad.test as a potential backend
}

