% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{adTestStatistic}
\alias{adTestStatistic}
\title{Anderson-Darling Test Statistic}
\usage{
adTestStatistic(y, z)
}
\arguments{
\item{y}{Outcome variable}

\item{z}{Treatment indicator vector, must be binary or logical}
}
\value{
Function to calculate AD test statistic
}
\description{
Calculates Anderson-Darling (AD) test statistic
}
\seealso{
\code{\link{RItest}}
}
