% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{aggregateDesigns}
\alias{aggregateDesigns}
\title{Aggregate DesignOptions}
\usage{
aggregateDesigns(design)
}
\arguments{
\item{design}{DesignOptions}
}
\value{
Aggregates
}
\description{
totals up all the covariates
}
