% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{alignDesignsByStrata}
\alias{alignDesignsByStrata}
\title{Align DesignOptions by Strata}
\usage{
alignDesignsByStrata(design, post.align.transform = NULL)
}
\arguments{
\item{design}{DesignOptions}

\item{post.align.transform}{A post-align transform}
}
\value{
list List of `CovsAlignedToADesign` objects
}
\description{
Align DesignOptions by Strata
}
\keyword{internal}
