% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\docType{class}
\name{AlignedCovs-class}
\alias{AlignedCovs-class}
\title{AlignedCovs S4 class}
\description{
A class for representing covariate matrices after alignment within stratum,
for a (single) given stratifying factor.  There can also be a clustering variable,
assumed to be nested within the stratifying variable. Extends DesignMatrix class.
}
\details{
Unlike DesignOptions this class has no UnitWeights slot.  Rather, the NotMissing
slot now records products of unit weights and non-missingness (more specifically
cluster means of non-missingness averaged with element weights, if provided). 
The StrataWeightRatio slot has an entry for each unit, representing ratio of
specified stratum weight to the product of h_b (the harmonic mean of n_{tb} and
n_{cb}, the counts of treatment and control clusters in stratum b) with bar-w_b,
(the arithmetic mean of aggregated cluster weights within that stratum).
}
\section{Slots}{

\describe{
\item{\code{Z}}{Logical indicating treatment assignment}

\item{\code{StrataMatrix}}{A sparse matrix with n rows and s columns, with 1 if the unit is in that stratification}

\item{\code{StrataFactor}}{Factor indicating strata}

\item{\code{StrataWeightRatio}}{For each unit, ratio of stratum weight to h_b; but see Details.}

\item{\code{Cluster}}{Factor indicating who's in the same cluster with who}

\item{\code{OriginalVariables}}{Look up table associating Covariates cols to terms in the calling formula, as in DesignMatrix}

\item{\code{Covariates}}{Numeric matrix, as in DesignMatrix, except here we presume columns to have been aligned (stratum-centered)}

\item{\code{NotMissing}}{matrix of unit weights, normalized within stratum to have mean 1. (As otherwise, NAs are 0s)}
}}

