% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{alignedToInferentials}
\alias{alignedToInferentials}
\title{Align to Inferentials}
\usage{
alignedToInferentials(alignedcovs)
}
\arguments{
\item{alignedcovs}{A CovsAlignedToADesign object}
}
\value{
list
}
\description{
Align to Inferentials
}
\keyword{internal}
