% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{as.matrix.DesignMatrix}
\alias{as.matrix.DesignMatrix}
\title{convert a DesignMatrix to an R matrix class object}
\usage{
\method{as.matrix}{DesignMatrix}(x, ...)
}
\arguments{
\item{x}{a DesignMatrix class object}

\item{...}{not used}
}
\value{
An object of R class matrix
}
\description{
Convert a Design matrix to an ordinary R matrix
}
\author{
Ben B Hansen
}
