% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\docType{class}
\name{AsymptoticTestStatistic-class}
\alias{AsymptoticTestStatistic-class}
\title{S4 AsymptoticTestStatistic class}
\description{
S4 AsymptoticTestStatistic class
}
\section{Slots}{

\describe{
\item{\code{asymptotic}}{function}
}}

