% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributablefx.R
\name{attributable.effect.helpers}
\alias{attributable.effect.helpers}
\alias{additive.attributable.effect}
\alias{max.attributable.effects}
\alias{min.attributable.effects}
\title{Attribute Effects helper functions}
\usage{
additive.attributable.effect(y, z, b, a)

\method{max}{attributable.effects}(y, z)

\method{min}{attributable.effects}(y, z)
}
\arguments{
\item{y}{y}

\item{z}{z}

\item{b}{b}

\item{a}{a}
}
\value{
value
}
\description{
Attribute Effects helper functions
}
