% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributablefx.R
\name{attributableEffects}
\alias{attributableEffects}
\title{Attributed Effects}
\usage{
attributableEffects(data, treatment, test.stat, samples = 5000,
  min.a = min.attributable.effects(data, treatment),
  max.a = max.attributable.effects(data, treatment), step = 1)
}
\arguments{
\item{data}{data}

\item{treatment}{treatment}

\item{test.stat}{test.stat}

\item{samples}{samples}

\item{min.a}{min.a}

\item{max.a}{max.a}

\item{step}{step}
}
\value{
RITest
}
\description{
Attributed Effects
}

