% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanceTestEngine.R
\name{balanceTestEngine}
\alias{balanceTestEngine}
\title{xBalance helper function}
\usage{
balanceTestEngine(ss, zz, mm, report, swt, s.p, normalize.weights, zzname,
  post.align.trans, p.adjust.method)
}
\arguments{
\item{ss}{ss}

\item{zz}{zz}

\item{mm}{mm}

\item{report}{report}

\item{swt}{swt}

\item{s.p}{s.p}

\item{normalize.weights}{normalize.weights}

\item{zzname}{zzname}

\item{post.align.trans}{post.align.trans}

\item{p.adjust.method}{Method to adjust P.}
}
\value{
List
}
\description{
Make engine
}

