% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{clusterRandomSampler}
\alias{clusterRandomSampler}
\title{Cluster Random Sampler}
\usage{
clusterRandomSampler(clusters, z)
}
\arguments{
\item{clusters}{Cluster membership vector}

\item{z}{Treatment assignment vector, must be binary or logical}
}
\value{
Function to use in \code{sampler} argument of \code{RItest}
}
\description{
Creates a function that draws from the set of all randomizations, assuming
treatment assignment to all observations within a cluster
}
\examples{
clusters <-  sample(letters[1:3], 10, replace = TRUE)
treatment <- ifelse(clusters == sample(letters[1:3], 1), 1, 0)
clustered <- clusterRandomSampler(clusters = clusters, z = treatment)
}
\seealso{
\code{\link{RItest}}
}
