% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsofeffect.R
\name{constant.multiplicative.model}
\alias{constant.multiplicative.model}
\title{Helper function}
\usage{
constant.multiplicative.model(ys, z, b, beta)
}
\arguments{
\item{ys}{the data to be adjusted, these methods assume a vector}

\item{z}{a numeric indicator of treament (0,1)}

\item{b}{a blocking indicator (may be a single level)}

\item{beta}{beta}
}
\value{
value
}
\description{
Helper function
}
