% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_SparseM_reduction_matrix}
\alias{create_SparseM_reduction_matrix}
\title{Helper function to slm.fit.csr.fixed}
\usage{
create_SparseM_reduction_matrix(zeroes)
}
\arguments{
\item{x}{logical vector indicating which entries of x'x are zeroes.}
}
\value{
SparseM matrix that will reduce the dimension of x'x and xy
}
\description{
This function generates a matrix that can be used to reduce
the dimensions of x'x and xy such that positive definiteness is
ensured and more practically, that SparseM::chol will work
}
