% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{cvmBackEnd}
\alias{cvmBackEnd}
\alias{.cvmBackEnd}
\title{cvm BackEnd}
\usage{
.cvmBackEnd(adjusted.y, z)
}
\arguments{
\item{adjusted.y}{y}

\item{z}{z}
}
\value{
RandomizationDistribution
}
\description{
this borrowed from cvm.test
}
