% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{cvmTestStatistic}
\alias{cvmTestStatistic}
\title{CVM Test Statistic}
\usage{
cvmTestStatistic(y, z)
}
\arguments{
\item{y}{Outcome Variable}

\item{z}{Treatment indicator vector, must be binary or logical}
}
\value{
Function to calculate CVM test statistic
}
\description{
Calculates Cramer-Von Mises Test Statistic
}
