% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{ddstTestStatistic}
\alias{ddstTestStatistic}
\title{ddst test Statistic}
\usage{
ddstTestStatistic(binn = 100, location = "median")
}
\arguments{
\item{binn}{binn}

\item{location}{location}
}
\value{
function
}
\description{
ddst test Statistic
}
