% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\docType{class}
\name{Design-class}
\alias{Design-class}
\title{Design S4 class}
\description{
Design S4 class
}
\section{Slots}{

\describe{
\item{\code{Z}}{Z}

\item{\code{StrataMatrices}}{StrataMatrices}

\item{\code{StrataFrame}}{StrataFrame}

\item{\code{Cluster}}{Cluster}

\item{\code{OriginalVariables}}{OriginalVariables}

\item{\code{Covariates}}{Covariates}

\item{\code{NotMissing}}{NotMissing}
}}

