% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{design_matrix}
\alias{design_matrix}
\title{Model matrices along with compact encodings of data availability/missingness}
\usage{
design_matrix(object, data = environment(object), remove.intercept = TRUE,
  ...)
}
\arguments{
\item{object}{Model formula or terms object (as in `model.matrix`)}

\item{data}{Data frame (as in `model.matrix`)}

\item{remove.intercept}{logical}

\item{...}{passed to `model.matrix.default` (and further)}
}
\value{
DesignMatrix instance of an S4 class that enriches model matrices with missing data info
}
\description{
Grow a model matrix while at the same time compactly
encoding missingness patterns in RHS variables of a model frame.
}
\author{
Ben B Hansen
}
