% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\docType{class}
\name{DesignMatrix-class}
\alias{DesignMatrix-class}
\title{DesignMatrix S4 class}
\description{
If the Covariates matrix has an intercept, it will only be in the first column.
}
\section{Slots}{

\describe{
\item{\code{Covariates}}{The numeric matrix that `model.matrix` would have returned.}

\item{\code{OriginalVariables}}{look-up table associating Covariates columns with terms of the originating model formula}

\item{\code{term.labels}}{labels of terms of the originating model formula}

\item{\code{contrasts}}{Contrasts, a list of contrasts or NULL, as returned by `model.matrix.default`}

\item{\code{NotMissing}}{Matrix of numbers in [0,1] with as many rows as Covariates but only one more col than there are distinct covariate missingness patterns (at least 1, nothing missing). First col is entirely T or 1, like an intercept.}

\item{\code{NM.Covariates}}{integer look-up table mapping Covariates columns to columns of NotMissing.  (If nothing missing for that column, this is 1.)}

\item{\code{NM.terms}}{integer look-up table mapping term labels to  columns of NotMissing}
}}

