% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\docType{class}
\name{DesignOptions-class}
\alias{DesignOptions-class}
\title{DesignOptions S4 class}
\description{
Extends the DesignMatrix class
}
\section{Slots}{

\describe{
\item{\code{Z}}{Logical indicating treatment assignment}

\item{\code{StrataMatrices}}{This is a list of sparse matrices, each with n rows and s columns, with 1 if the unit is in that stratification}

\item{\code{StrataFrame}}{Factors indicating strata (not the sparse matrices, as we use them in the weighting function)}

\item{\code{Cluster}}{Factor indicating who's in the same cluster with who}

\item{\code{NotMissing}}{Matrix of numbers in [0,1], cluster means of element weights with missing values
contributing 0s.  First col is entirely T or 1, like an intercept, unless corresponding
ElementWeight is 0, in which case it may also be 0. Subsequent cols present only if
there are missing covariate values, in which case these cols are named for terms that
possess missing values.  Terms with the same missing data pattern are mapped to a single}
}}

