% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{DesignWeights}
\alias{DesignWeights}
\title{Create stratum weights to be associated with a DesignOptions}
\usage{
DesignWeights(design, stratum.weights = harmonic)
}
\arguments{
\item{design}{DesignOptions}

\item{stratum.weights}{Stratum weights}
}
\value{
list with two entries, \code{sweights}  and \code{wtratio}.
Each is in turn a list with an entry for each stratum.
The \code{sweights} will sum to 1.  \code{wtratio} is a constant multiple
of the ratio of the stratum weight s_b to h_b, the harmonic mean of n_{tb} and n_{cb},
the numbers of treatment and control clusters in stratum b. Not h_b times something
having to do with cluster sizes or weights, just h_b. the arbitrary constant multiple
out front is harmless b/c this is only used in inferential calcs; that constant will be
washed out en route to calculating z statistics and the d^2 stat.
}
\description{
Apply weighting function to a DesignOptions by stratum, returning results in a format
suitable to be associated with an existing design and used in further calcs.
}
\details{
(Developer note: There's no real reason not to simplify by only returning the sweights,
not also the wtratio's.  Weight ratios are used downstream in alignedToInferentials, but
the material in the harmonic means calcs is already being assembled there for other
reasons.  An improvement for another day...)
}
