% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{effectOfTreatmentOnTreated}
\alias{effectOfTreatmentOnTreated}
\title{Number of treatment clusters by stratum}
\usage{
effectOfTreatmentOnTreated(data)
}
\arguments{
\item{data}{Data.}
}
\value{
Cluster count
}
\description{
Calculate the number of treatment clusters by stratum --
these being proportional to "effect of treatment on treated" weights
when assignment probabilities are uniform within each stratum.
}
\details{
NB: currently, i.e. as of this inline note's commit,
this function is used only in testing.
}
\keyword{internal}
