% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{entropyRelDistTestStatistic}
\alias{entropyRelDistTestStatistic}
\title{entropy relative distance}
\usage{
entropyRelDistTestStatistic(smooth = 0.01, binn = 100, location = "median")
}
\arguments{
\item{smooth}{smooth}

\item{binn}{binn}

\item{location}{location}
}
\value{
function
}
\description{
entropy relative distance
}
