% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{farray}
\alias{farray}
\title{Fill in an array from a function that takes the indices as arguments.}
\usage{
farray(f, p)
}
\arguments{
\item{f}{The function to be called.}

\item{p}{The parameter list(a = c(1,2,3), b = c(7,8), c = c(9,10,11)) will create a 3x2x3 array.}
}
\value{
An array of the cartesian product of the parameters, with f applied to each combination.
}
\description{
Fill in an array from a function that takes the indices as arguments.
}
