% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten.xbalresult.R
\name{flatten.xbalresult}
\alias{flatten.xbalresult}
\title{Flattens xBalance output.}
\usage{
flatten.xbalresult(x, show.signif.stars = getOption("show.signif.stars"),
  show.pvals = !show.signif.stars, ...)
}
\arguments{
\item{x}{x}

\item{show.signif.stars}{Should signif stars be shown?}

\item{show.pvals}{Should p-vals be shown?}

\item{...}{Ignored}
}
\value{
Structure
}
\description{
Details...
}
