% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{formula.xbal}
\alias{formula.xbal}
\title{Returns \code{formula} attribute of an \code{xbal} object.}
\usage{
\method{formula}{xbal}(x, ...)
}
\arguments{
\item{x}{An \code{xbal} object.}

\item{...}{Ignored.}
}
\value{
The formula corresponding to \code{xbal}.
}
\description{
Returns \code{formula} attribute of an \code{xbal} object.
}
