% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\name{getApplyFunction}
\alias{getApplyFunction}
\title{Get proper apply function}
\usage{
getApplyFunction(x, fun, ...)
}
\arguments{
\item{x}{x}

\item{fun}{fun}

\item{...}{Add'l arguments.}
}
\value{
function
}
\description{
Get proper apply function
}
