% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\name{getLApplyFunction}
\alias{getLApplyFunction}
\title{Get proper lapply function}
\usage{
getLApplyFunction(x, fun, ...)
}
\arguments{
\item{x}{x}

\item{fun}{fun}

\item{...}{Add'l arguments.}
}
\value{
function
}
\description{
Get proper lapply function
}

