% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\docType{methods}
\name{givenParams,UniformityModel-method}
\alias{givenParams,UniformityModel-method}
\title{givenParams for UniformityModel}
\usage{
\S4method{givenParams}{UniformityModel}(model, ...)
}
\arguments{
\item{model}{Model}

\item{...}{Add'l arguments}
}
\value{
Output
}
\description{
givenParams for UniformityModel
}

