% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\name{givenParams}
\alias{givenParams}
\title{Generic function givenParms}
\usage{
givenParams(model, ...)
}
\arguments{
\item{model}{Model}

\item{...}{Add'l arguments}
}
\value{
Output
}
\description{
Generic function givenParms
}

