% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gramian_reduction}
\alias{gramian_reduction}
\title{Helper function to slm_fit_csr}
\usage{
gramian_reduction(zeroes)
}
\arguments{
\item{zeroes}{logical vector indicating which entries of the diagonal of x'x are zeroes.}
}
\value{
SparseM matrix that will reduce the dimension of x'x and xy
}
\description{
This function generates a matrix that can be used to reduce
the dimensions of x'x and xy such that positive definiteness is
ensured and more practically, that SparseM::chol will work
}
