% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{harmonic_times_mean_weight}
\alias{harmonic_times_mean_weight}
\title{Harmonic mean times mean of weights}
\usage{
harmonic_times_mean_weight(data)
}
\value{
numeric vector of length \code{nlevels(data$stratum.code)}
}
\description{
Harmonic mean times mean of weights
}
\keyword{internal}
