% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{HB08_2016}
\alias{HB08_2016}
\title{Hansen & Bowers (2008) inferentials 2016 [81e3ecf] version}
\usage{
HB08_2016(alignedcovs)
}
\arguments{
\item{alignedcovs}{A CovsAlignedToADesign object}
}
\value{
list, as in \code{\link{HB08}}
}
\description{
Hansen & Bowers (2008) inferentials 2016 [81e3ecf] version
}
\keyword{internal}
