% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{independentProbabilitySampler}
\alias{independentProbabilitySampler}
\title{Independent Probability Sampler}
\usage{
independentProbabilitySampler(n, p = rep(0.5, n))
}
\arguments{
\item{n}{n}

\item{p}{p}
}
\value{
List
}
\description{
Draw from the 2^n possible randomizations for n units with
probability vector p (e.g. where each unit gets its own, possibly
biased, coin flip)
}

