% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\name{invertModel}
\alias{invertModel}
\title{Invert a model}
\usage{
invertModel(m, y_0, z, ...)
}
\arguments{
\item{m}{Model}

\item{y_0}{Null}

\item{z}{z}

\item{...}{Add'l arguments}
}
\value{
Model inverse
}
\description{
apply the inverse of the model to the uniformity trial data to see
what we would observe given z and parameters
}

