% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipr.R
\name{ipr}
\alias{ipr}
\title{Inverse probability weighting suitable for completely randomized designs}
\usage{
ipr(z, strata, clus = NULL, type = "inverse probability")
}
\arguments{
\item{z}{categorical variable recording distinctions among assignments to treatment conditions}

\item{strata}{categorical variable recording blocks}

\item{clus}{optional categorical variable indicating cluster membership}

\item{type}{character string naming desired type of weight. See Details for options other than inverse probability}
}
\value{
vector of weights
}
\description{
Assemble inverse probability of assignment weights on the fly,
e.g. as part of evaluation of `weights` argument to a `glm` or similar
function.   Also calculates inverse odds of assignment weights.
Assumes a design that is completely randomized, either by cluster or by
the unit of observation, potentially within strata or blocks.
}
\details{
The function assumes complete random assignment, within (optional)
blocks.  This can be complete random assignment of clusters; in this situation
the user avoids having to manage aggregation to the cluster level in order
properly to infer treatment probabilities.  There need to be at least two conditions,
and any block without all conditions represented will be dropped, ie given observation 
weights of 0, as are observations that are NA on the blocking variable.

NAs in the assignment variable also translate to zero.  If they occur within a
non-NA block, probabilities or odds for that block are calculated as though the 
observations with NAs for \code{z} just weren't there.
}
\author{
Ben B. Hansen
}

