% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{ksBackEnd}
\alias{.ksBackEnd}
\alias{ksBackEnd}
\title{this borrowed from ks.test}
\usage{
.ksBackEnd(adjusted.y, z)
}
\arguments{
\item{adjusted.y}{y}

\item{z}{z}
}
\value{
RandomizationDistribution
}
\description{
this borrowed from ks.test
}

