% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{ksTestStatistic}
\alias{ksTestStatistic}
\title{KS Test Statistic}
\usage{
ksTestStatistic(y, z)
}
\arguments{
\item{y}{Outcome variable}

\item{z}{Treatment indicator vector, must be binary or logical}
}
\value{
Function to calculate KS test statistic
}
\description{
Calculates Kolmogorov-Smirnov test statistic
}
\seealso{
\code{\link{RItest}}
}
