% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{madDiff}
\alias{madDiff}
\title{Median Absolute Deviation}
\usage{
madDiff(ys, z)
}
\arguments{
\item{ys}{Outcome variable}

\item{z}{Treament indicator vector}
}
\value{
Median absolute deviation
}
\description{
Evaluates median of the absolutes deviations from the median.
}
\seealso{
\code{\link{RItest}} \code{\link[stats]{mad}}
}
