% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{makeDesign}
\alias{makeDesign}
\title{Create a Design object from a formula and some data}
\usage{
makeDesign(fmla, data, imputefn = median, na.rm = FALSE,
  include.NA.flags = TRUE)
}
\arguments{
\item{fmla}{Formula}

\item{data}{Data}

\item{imputefn}{Function to impute}

\item{na.rm}{Should NA's removed?}

\item{include.NA.flags}{Flag NA's?}
}
\value{
Design
}
\description{
The formula must have a left hand side that can be converted to a logical.
}
\details{
On the RHS:
 - It may have at most one cluster() argument.
 - It may have one or more strata() arguments.
 - All other variables are considered covariates.

NAs in a cluster() or strata() variable will be dropped.  NAs in covariates will be imputed.
}
