% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{makeDesigns}
\alias{makeDesigns}
\title{Create a DesignOptions object from a formula and some data}
\usage{
makeDesigns(fmla, data)
}
\arguments{
\item{fmla}{Formula}

\item{data}{Data}
}
\value{
DesignOptions
}
\description{
The formula must have a left hand side that can be converted to a logical.
}
\details{
On the RHS:
 - It may have at most one cluster() argument.
 - It may have one or more strata() arguments.
 - All other variables are considered covariates.

NAs in a cluster() or strata() variable will be dropped.
NAs in covariates will be passed through, but without
being flagged as NotMissing (as available data items will)
}
