% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelsofeffect.R
\name{min.max.model}
\alias{min.max.model}
\title{Min Max Model}
\usage{
\method{min}{max.model}(model, lower = -Inf, upper = Inf)
}
\arguments{
\item{model}{Model}

\item{lower}{Lower bound}

\item{upper}{Upper bound}
}
\value{
New model
}
\description{
Given lower and lower bounds a and b, create a new model that
limits within those bounds
}

