% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\name{model.effect.helper}
\alias{model.effect.helper}
\title{Helper function for model.effect}
\usage{
model.effect.helper(model, combine, R, Z, ...)
}
\arguments{
\item{model}{model}

\item{combine}{combine function}

\item{R}{R}

\item{Z}{Z}

\item{...}{Add'l arguments.}
}
\value{
Result
}
\description{
Helper function for model.effect
}
