% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{model_matrix}
\alias{model_matrix}
\title{Model matrices along with compact encodings of data availability/missingness}
\usage{
model_matrix(object, data = environment(object),
  remove.intercept = TRUE, ...)
}
\arguments{
\item{object}{Model formula or terms object (as in `model.matrix`)}

\item{data}{data.frame, as in `model.matrix()` but has to have \sQuote{\code{(weights)}} column}

\item{remove.intercept}{logical}

\item{...}{passed to `model.matrix.default` (and further)}
}
\value{
ModelMatrixPlus, i.e. model matrix enriched with missing data info
}
\description{
Grow a model matrix while at the same time compactly
encoding missingness patterns in RHS variables of a model frame.
}
\author{
Ben B Hansen
}
\keyword{internal}
