% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{odds.ratio}
\alias{odds.ratio}
\title{Odds ratio}
\usage{
odds.ratio(y, z)
}
\arguments{
\item{y}{Outcome variable, must be binary or logical}

\item{z}{Treatment indicator vector, must be binary or logical}
}
\value{
Function to calculate odds ratio
}
\description{
Calculates odds ratio
}
\seealso{
\code{\link{RItest}}
}
