% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.xbal.R
\name{original_units_var_formatter}
\alias{original_units_var_formatter}
\title{Formatting suitable for stat expressed in units specific to var}
\usage{
original_units_var_formatter(arr, digits, var_format = list())
}
\arguments{
\item{arr}{numeric array}

\item{digits}{number of digits for rounding}

\item{var_format}{A list of lists. Each named item of the outer list will be matched to a variable. The inner lists should have two items, `mean` and `diff`. The first formats statistics based on averages. The `diff` item should format statistics that are differences.}
}
\value{
array of same dimension as arr but of type character
}
\description{
formats a var-stat-strata array by var, with rounding
potentially rounding a bit less for an "adj.diff" column
}
\author{
Hansen
}
\keyword{internal}
