% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzpvalues.R
\name{p-value}
\alias{general.two.sided.mid.p.value}
\alias{general.two.sided.p.value}
\alias{lower.p.value}
\alias{p-value}
\alias{simple.p.value}
\alias{upper.p.value}
\title{P-value gnerators}
\usage{
simple.p.value(value, distribution, lower.tail = FALSE)

upper.p.value(value, distribution)

lower.p.value(value, distribution)

general.two.sided.mid.p.value(value, distribution)

general.two.sided.p.value(value, distribution)
}
\arguments{
\item{value}{value}

\item{distribution}{distribution}

\item{lower.tail}{lower.tail?}
}
\value{
value
}
\description{
P-value gnerators
}

