% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzpvalues.R
\name{p.values}
\alias{p.values}
\title{Generate p-values}
\usage{
p.values(object, p.value.function = general.two.sided.p.value)
}
\arguments{
\item{object}{Object}

\item{p.value.function}{Function to use}
}
\value{
ParameterPvals
}
\description{
Generate p-values
}
