% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\name{parallelLoaded}
\alias{parallelLoaded}
\title{Helper function to check if parallel package is loaded}
\usage{
parallelLoaded()
}
\value{
Logical
}
\description{
Helper function to check if parallel package is loaded
}
