% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\docType{class}
\name{ParameterizedTest-class}
\alias{ParameterizedTest-class}
\title{the RandomizationDistribution object}
\description{
for a single test statistic, results of a call to
randomizationDistribution, each distribution is with respect to a
given test statistic
}
\section{Slots}{

\describe{
\item{\code{sharp.null}}{SharpNullTest object}
}}

