% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDistance.R
\name{parameterSensitivity}
\alias{parameterSensitivity}
\title{Compares the predictions of a model over the parameter space.}
\usage{
parameterSensitivity(model, parameters, uniformity, z)
}
\arguments{
\item{model}{A \code{\link{UniformityModel}} object that will be
used with the parameters to generate the model predictions.}

\item{parameters}{A list of the form \code{list(param1 = c(1,2,3),
param2 = c(4,5,6), ...)} where the names are the parameter
arguments to the model function.}

\item{uniformity}{Data that represent the uniformity trial (the
state of the world when treatment is not applied to any unit).}

\item{z}{A vector of treatment indicators.}
}
\value{
A matrix of two columns: the parameter space distance and
  the prediction distance for all pair-wise comparisons
}
\description{
Models that are insensitive to changes in the parameters will have
little distance between predictions.
}

