% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDistance.R
\name{parameterSensitivityBreakoutPlot}
\alias{parameterSensitivityBreakoutPlot}
\title{Plots the results of the parameter sensitivity on the original parameter space.}
\usage{
parameterSensitivityBreakoutPlot(x, subset, col = rgb(0, 0, 0, 0.25),
  ...)
}
\arguments{
\item{x}{The results of the call to \code{\link{parameterSensitivity}}.}

\item{subset}{An expression that will evaluate to a logical value when
evaluated in the context of \code{x}. You can use \code{parameter} and
\code{prediction} to subset which parameters to show.}

\item{col}{The color of the lines, including transparency.}

\item{...}{Other parameters to pass to \code{\link{segments}}.}
}
\description{
Works best for 2D models, may not work for other dimenions.
}
