% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\name{plot.ParameterizedTest}
\alias{plot.ParameterizedTest}
\title{Plot the results of RItest.}
\usage{
\method{plot}{ParameterizedTest}(x, type = "o", summary = max, ...)
}
\arguments{
\item{x}{The results of RItest.}

\item{type}{The type (points, lines, etc) for one dimensional
models. Ignored for multi-dimensional models.}

\item{summary}{A function to summarize higher dimension models (p >
2) into 2 dimensions.}

\item{...}{Add'l arguments}
}
\value{
A lattice object. Use print to plot.
}
\description{
Plot the results of RItest.
}

