% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzpvalues.R
\name{plot.ParameterPvals}
\alias{plot.ParameterPvals}
\title{Plot function for ParameterPvals}
\usage{
\method{plot}{ParameterPvals}(object, ...)
}
\arguments{
\item{object}{ParameterPvals}

\item{...}{Add'l arguments to \code{plot}}
}
\value{
plot
}
\description{
Plot function for ParameterPvals
}

