% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelDistance.R
\name{plot.parameterSensitivity}
\alias{plot.parameterSensitivity}
\title{Plots the results of \code{\link{parameterSensitivity}} as a 2D histogram.}
\usage{
\method{plot}{parameterSensitivity}(x, ...)
}
\arguments{
\item{x}{The matrix of distances from \code{\link{parameterSensitivity}}.}

\item{...}{Arguments to be passed to the plot method for \code{\link{hexbin}}}
}
\description{
Plots the results of \code{\link{parameterSensitivity}} as a 2D histogram.
}
