% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\docType{methods}
\name{point,ParameterizedTest-method}
\alias{point,ParameterizedTest-method}
\title{point for ParameterizedTest}
\usage{
\S4method{point}{ParameterizedTest}(object)
}
\arguments{
\item{object}{ParameterizedTest}
}
\value{
output
}
\description{
point for ParameterizedTest
}

