% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnitude.R
\name{print.balanceMagnitude}
\alias{print.balanceMagnitude}
\title{Print a magnitude balance report}
\usage{
\method{print}{balanceMagnitude}(x, digits = 3, ...)
}
\arguments{
\item{x}{a \code{balanceMagnitude} object}

\item{digits}{number of digits for the standardized differences}

\item{...}{ignored}
}
\value{
\code{x}, invisibly
}
\description{
Print a magnitude balance report
}
