% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{quantileAbsoluteDifference}
\alias{quantileAbsoluteDifference}
\title{Quantile Differences}
\usage{
quantileAbsoluteDifference(quantiles)
}
\arguments{
\item{quantiles}{quantiles}
}
\value{
value
}
\description{
Similar to the KS test -- but evaluates differences at discrete
points of the eCDF
}

