% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{quantileDifference}
\alias{quantileDifference}
\title{Quantile difference}
\usage{
quantileDifference(q = 0.5)
}
\arguments{
\item{q}{quantiles}
}
\value{
diff
}
\description{
Quantile difference
}

