% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{quasireldist}
\alias{quasireldist}
\title{Quasi Relative Distance}
\usage{
quasireldist(y, yo, binn = 100, location = "median")
}
\arguments{
\item{y}{y}

\item{yo}{y0}

\item{binn}{binn}

\item{location}{location}
}
\value{
vector
}
\description{
Quasi Relative Distance
}

