% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzpvalues.R
\docType{class}
\name{RandomizationDistribution-class}
\alias{RandomizationDistribution-class}
\title{RandomizationDistribution}
\description{
RandomizationDistribution
}
\section{Slots}{

\describe{
\item{\code{test.statistic}}{xBalance object}
}}

