% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzpvalues.R
\docType{class}
\name{RandomizationDistributionConfInt-class}
\alias{RandomizationDistributionConfInt-class}
\title{Randomization Distribution Confidence Interval}
\description{
Randomization Distribution Confidence Interval
}
\section{Slots}{

\describe{
\item{\code{prd}}{RandomizationDistribution}

\item{\code{rejections}}{data.frame}
}}

