% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{rank.sum.functions}
\alias{paired.sgnrank.sum}
\alias{rank.sum}
\alias{rank.sum.functions}
\title{Rank based functions}
\usage{
rank.sum(ys, z)

paired.sgnrank.sum(ys, z)
}
\arguments{
\item{ys}{ys}

\item{z}{Vector that can be made logical}
}
\value{
sum
}
\description{
Rank based functions
}

