% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\name{returnTau}
\alias{returnTau}
\title{helper function for all the models that have a "tau" direct effect}
\usage{
returnTau(...)
}
\arguments{
\item{...}{Arguments including tau}
}
\value{
tau
}
\description{
helper function for all the models that have a "tau" direct effect
}

