% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\name{RItest}
\alias{RItest}
\title{Test a sharp hypothesis}
\usage{
RItest(y, z, test.stat, moe = NULL, parameters = NULL,
  sampler = simpleRandomSampler(z = z, b = rep(1, length(z))),
  samples = 5000, type = "exact", ...)
}
\arguments{
\item{y}{Outcome variable of interest}

\item{z}{Treatment indicator vector, must be binary or logical}

\item{test.stat}{Test statistic to be reported}

\item{moe}{Model of effect, single function with signature 
f(y, z, param1, param2, etc.)}

\item{parameters}{Parameters values in model of effect 
List of name = values, name = values, ...}

\item{sampler}{Function specifying sampling procedure, uses simple 
randomization by default}

\item{samples}{Number of samples to use in direct permutation, uses 5000 
by default}

\item{type}{Either \code{"exact"} (direct permutation) or 
\code{"asymptotic"} (distributional approximation). Currently, 
\code{type = "asymptotic"} is not supported}

\item{...}{Additional arguments}
}
\value{
Observed test statistic and associated p-value
}
\description{
Test a sharp null hypothesis based on a experimental design.
Computes a test statistic of choice and reports p-values using direct
permutation
}
\examples{
data(nuclearplants)
RItest(nuclearplants$cost, nuclearplants$pr, test.stat = t.mean.difference)
}
\seealso{
Update with vignettes here
}
