\name{RItools-internal}
\alias{RItools-internal}

\title{Helper functions temporarily not worth documenting.}
\alias{comparesimp.to.pair.xb.fn}
\alias{bigchoose}
\alias{mean.difference}
\alias{rank.sum}
\alias{constant.additive.hypothesis.factory}
\alias{constant.hypotheses}
\alias{lm.residualizer}
\alias{randomization.distribution}
\alias{expand.z}
\alias{simple.p.value}
\alias{manual.symmetric.p.value}
\alias{auto.symmetric.p.value}
\alias{general.two.sided.mid.p.value}
\alias{general.two.sided.p.value}
\alias{confint}
\alias{summary.confint}
\alias{rd.histograms}	
       
\description{Functions not worth documenting}
\keyword{ internal }
