% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\docType{class}
\name{SharpNullTest-class}
\alias{SharpNullTest-class}
\title{SharpNullTest class}
\description{
SharpNullTest class
}
\section{Slots}{

\describe{
\item{\code{observed.statistic}}{observed.statistic}

\item{\code{samples}}{sample}

\item{\code{call}}{call}
}}

