% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{slm.fit.csr.fixed}
\alias{slm.fit.csr.fixed}
\title{slm.fit.csr with a fix}
\usage{
slm.fit.csr.fixed(x, y, ...)
}
\arguments{
\item{x}{As slm.fit.csr}

\item{y}{As slm.fit.csr}

\item{...}{As slm.fit.csr}
}
\value{
As slm.fit.csr
}
\description{
SparseM's slm.fit.csr has a bug for intercept only models
(admittedly, these are generally a little silly to be done as a
sparse matrix), but in order to avoid duplicate code, if
everything is in a single strata, we use the intercept only model.
}

