% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slm.wfit.csr}
\alias{slm.wfit.csr}
\title{slm.wfit with two fixes}
\usage{
slm.wfit.csr(x, y, weights, ...)
}
\arguments{
\item{x}{As slm.wfit}

\item{y}{As slm.wfit}

\item{weights}{As slm.wfit}

\item{...}{As slm.wfit}
}
\value{
As slm.wfit
}
\description{
slm.wfit shares the intercept-only issue with slm.fit,
and in addition has an issue where it carries forward
incorrect residuals and fitted values.
}
