% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\name{snowLoaded}
\alias{snowLoaded}
\title{Helper function to check if snow package is loaded}
\usage{
snowLoaded()
}
\value{
Logical
}
\description{
Helper function to check if snow package is loaded
}

