% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SparseMMFromFactor}
\alias{SparseMMFromFactor}
\title{Sparse matrix dummy coding of a factor variable (omitting the intercept)}
\usage{
SparseMMFromFactor(thefactor)
}
\arguments{
\item{thefactor}{Factor variable, or object inheriting from class factor}
}
\value{
Sparse csr matrix the columns of which are dummy variables for levels of thefactor
}
\description{
Turn a factor variable into a sparse matrix of 0's and 1's, such that if observation i
has the jth level then there is a 1 at position (i,j) (but nowhere else in row i).
}
\details{
NA's give rise to rows with no 1s.
As the result is only meaningful in the context of the SparseM package,
function requires that SparseM be loaded.
}
\examples{
sparse_mod_matrix <-  SparseMMFromFactor(iris$Species)
mod_matrix <- model.matrix(~Species-1, iris)
all.equal(as.matrix(sparse_mod_matrix),
          mod_matrix, check.attributes=FALSE)
}
\author{
Ben Hansen
}

