% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{sparseToVec}
\alias{sparseToVec}
\title{Convert Matrix to vector}
\usage{
sparseToVec(s, column = TRUE)
}
\arguments{
\item{s}{Matrix}

\item{column}{Column (TRUE) or row (FALSE)?}
}
\value{
vector
}
\description{
Convert Matrix to vector
}
