% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{ssrBackEnd}
\alias{ssrBackEnd}
\alias{.ssrBackEnd}
\title{ssr backEnd}
\usage{
.ssrBackEnd(adjusted.y, z, d)
}
\arguments{
\item{adjusted.y}{y}

\item{z}{z}

\item{d}{d}
}
\value{
RandomizationDistribution
}
\description{
ssr backEnd
}
