% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{ssrTSmaker}
\alias{ssrTSmaker}
\title{ssrTestStatistic}
\usage{
ssrTSmaker(S)
}
\arguments{
\item{S}{S}
}
\value{
AsymptoticTestStatistic
}
\description{
The ssrTestStatistic with the F-test as a potential asymp backend
(and using F as the test statistic)
}

