% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\docType{class}
\name{StratumWeightedDesignOptions-class}
\alias{StratumWeightedDesignOptions-class}
\title{Stratum Weighted DesignOptions}
\description{
Stratum Weighted DesignOptions
}
\section{Slots}{

\describe{
\item{\code{Sweights}}{stratum weights}
}}

