% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subset.xbal}
\alias{subset.xbal}
\title{Select variables, strata, and statistics from a \code{xbal} object}
\usage{
\method{subset}{xbal}(x, vars = NULL, strata = NULL, stats = NULL,
  tests = NULL, ...)
}
\arguments{
\item{x}{The \code{xbal} object, the result of a call to
\code{\link{xBalance}}}

\item{vars}{The variable names to select.}

\item{strata}{The strata names to select.}

\item{stats}{The names of the variable level statistics to select.}

\item{tests}{The names of the group level tests to select.}

\item{...}{Other arguments (ignored)}
}
\value{
A \code{xbal} object with just the appropriate items
  selected.
}
\description{
If any of the arguments are not specified, all the of relevant
items are included.
}
