% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{subsetStatistic}
\alias{subsetStatistic}
\title{Subset / subgroup Analysis}
\usage{
subsetStatistic(statistic, ...)
}
\arguments{
\item{statistic}{statistic}

\item{...}{Add'l arguments to subset}
}
\value{
function
}
\description{
using a logical vector, limit the test statistic to a smaller group
}

