% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{test.stat.means}
\alias{test.stat.means}
\alias{fastmean}
\alias{mean.difference}
\alias{mean.diff.lsfit}
\alias{mean.diff.vect}
\alias{t.mean.difference}
\alias{fastvar}
\title{Mean functions}
\usage{
fastmean(ys)

\method{mean}{difference}(ys, z)

\method{mean}{diff.lsfit}(ys, z)

\method{mean}{diff.vect}(ys, z)

t.mean.difference(ys, z)

fastvar(ys)
}
\arguments{
\item{ys}{Outcome variable}

\item{z}{Treatment indicator vector that can be made logical}
}
\value{
Function to calculate mean
}
\description{
Various functions to compute means
}
\seealso{
\code{\link{RItest}}
}
