% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\name{testSize}
\alias{testSize}
\title{testSize}
\usage{
testSize(analysis, alpha = NULL, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{analysis}{analysis}

\item{alpha}{alpha}

\item{tol}{tol}
}
\value{
size of test
}
\description{
given the results of analyzeModel, what can we say?
}

