% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\docType{class}
\name{UniformityModel-class}
\alias{UniformityModel-class}
\title{The most generic type of model is a "UniformityModel" which is a
function that takes y and z (and possibly parameters) and returns
the uniformity trial the @inverse slot contains a function that
given the uniformity trial data, z, and params returns the
observed data implied by the model.}
\description{
The most generic type of model is a "UniformityModel" which is a
function that takes y and z (and possibly parameters) and returns
the uniformity trial the @inverse slot contains a function that
given the uniformity trial data, z, and params returns the
observed data implied by the model.
}
\section{Slots}{

\describe{
\item{\code{inverse}}{inverse}
}}

