% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizationModels.R
\name{UniformityModel}
\alias{UniformityModel}
\title{Create new UniformityModel}
\usage{
UniformityModel(f, inv)
}
\arguments{
\item{f}{Function}

\item{inv}{Inverse}
}
\value{
UniformityModel
}
\description{
Create new UniformityModel
}

