% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{weightedDesign}
\alias{weightedDesign}
\title{Create a weighted Design}
\usage{
weightedDesign(design, stratum.weights = harmonic, normalize.weights = TRUE)
}
\arguments{
\item{design}{Design}

\item{stratum.weights}{Stratum weights}

\item{normalize.weights}{Normalize weights?}
}
\value{
WeightedDesign
}
\description{
Adds weights to existing Design.
}

