% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{wilcoxBackEnd}
\alias{wilcoxBackEnd}
\alias{.wilcoxBackEnd}
\title{Wilcox BackEnd}
\usage{
.wilcoxBackEnd(adjusted.data, treatment)
}
\arguments{
\item{adjusted.data}{adjusted.data}

\item{treatment}{treatment}
}
\value{
RandomizationDistribution
}
\description{
Wilcox BackEnd
}
