% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{withOptions}
\alias{withOptions}
\title{Safe way to temporarily override options()}
\usage{
withOptions(optionsToChange, fun)
}
\arguments{
\item{optionsToChange}{Which options.}

\item{fun}{Function to run with new options.}
}
\value{
Result of \code{fun}.
}
\description{
Safe way to temporarily override options()
}

