% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBalance.find.goodstrats.R
\name{xBalance.find.goodstrats}
\alias{xBalance.find.goodstrats}
\title{xBalance helper function}
\usage{
xBalance.find.goodstrats(ss.df, zz, mm)
}
\arguments{
\item{ss.df}{Degrees of freedom.}

\item{zz}{Treatment}

\item{mm}{mm}
}
\value{
Data.frame
}
\description{
Finds good strata
}

