% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBalance.make.stratwts.R
\name{xBalance.make.stratwts}
\alias{xBalance.make.stratwts}
\title{xBalance helper function}
\usage{
xBalance.make.stratwts(stratum.weights, ss.df, zz, data, normalize.weights)
}
\arguments{
\item{stratum.weights}{Weights}

\item{ss.df}{df.}

\item{zz}{treatment}

\item{data}{data}

\item{normalize.weights}{weights}
}
\value{
list
}
\description{
Makes strata weights
}

