% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBalance.makeMM.R
\name{xBalance.makeMM}
\alias{xBalance.makeMM}
\title{Helper function for xBalance}
\usage{
xBalance.makeMM(tfm, dat)
}
\arguments{
\item{tfm}{tfm}

\item{dat}{data}
}
\value{
mm
}
\description{
Make MM
}

