% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBalance.makepooledsd.R
\name{xBalance.makepooledsd}
\alias{xBalance.makepooledsd}
\title{xBalance helper function}
\usage{
xBalance.makepooledsd(zz, mm, pre.n)
}
\arguments{
\item{zz}{Treatment}

\item{mm}{mm}

\item{pre.n}{pre.n}
}
\value{
pooled SD
}
\description{
Make pooled SD
}
