% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{xBalanceBackEnd}
\alias{.xBalanceBackEnd}
\alias{xBalanceBackEnd}
\title{xBalanceBackEnd}
\usage{
.xBalanceBackEnd(adjusted.data, treatment)
}
\arguments{
\item{adjusted.data}{adjusted.data}

\item{treatment}{treatment}
}
\value{
RandomizationDistribution
}
\description{
xBalanceBackEnd
}

