% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xBalanceEngine.R
\name{xBalanceEngine}
\alias{xBalanceEngine}
\title{xBalance helper function}
\usage{
xBalanceEngine(
  ss,
  zz,
  mm,
  report,
  swt,
  s.p,
  normalize.weights,
  zzname,
  post.align.trans
)
}
\arguments{
\item{ss}{ss}

\item{zz}{zz}

\item{mm}{mm}

\item{report}{report}

\item{swt}{swt}

\item{s.p}{s.p}

\item{normalize.weights}{normalize.weights}

\item{zzname}{zzname}

\item{post.align.trans}{post.align.trans}
}
\value{
List
}
\description{
Make engine
}
