% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{aggregateDesigns}
\alias{aggregateDesigns}
\title{Aggregate DesignOptions}
\usage{
aggregateDesigns(design)
}
\arguments{
\item{design}{DesignOptions}
}
\value{
another DesignOptions representing the clusters
}
\description{
Totals up all the covariates, as well as user-provided unit weights.
(What it does to NotMissing entries is described in docs for DesignOptions class.)
}
\details{
If \code{design@Cluster} has extraneous (non-represented) levels, they will be dropped.
}
\keyword{internal}
