% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{alignDesignsByStrata}
\alias{alignDesignsByStrata}
\title{Align DesignOptions by Strata}
\usage{
alignDesignsByStrata(a_stratification, design,
  post.align.transform = NULL)
}
\arguments{
\item{a_stratification}{name of a column of `design@strataFrame`/ element of `design@Sweights`}

\item{design}{DesignOptions}

\item{post.align.transform}{A post-align transform (cf \code{\link{balanceTest}})}
}
\value{
CovsAlignedToADesign
}
\description{
Align DesignOptions by Strata
}
\keyword{internal}
