% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\docType{class}
\name{AlignedCovs-class}
\alias{AlignedCovs-class}
\title{AlignedCovs S4 class}
\description{
A class for representing covariate matrices after alignment within stratum,
for a (single) given stratifying factor.  There can also be a clustering variable,
assumed to be nested within the stratifying variable. Extends DesignMatrix.
}
\details{
Just as the covariates are presumed to have been aligned, the NotMissing weights carried
in this class are presumed to have been normalized: in each stratum, either all
weights are 0, or they've been rescaled to sum to 1. In order to handle different
NA patterns for different covariates, this normalization is done separately for each variable.

In addition, unlike DesignOptions this class has no ElementWeights slot.
So the burden of representing differences in stratum
sizes falls entirely on the StrataWeightRatio slot. This slot has an entry for each unit,
representing ratio of user provided or specified stratum weight to h_b, the harmonic
mean of n_{tb} and n_{cb}, the counts of treatment and control clusters in stratum b.
(So h_b in no way reflects element weights or cluster sizes.)
}
\section{Slots}{

\describe{
\item{\code{Z}}{Logical indicating treatment assignment}

\item{\code{StrataMatrix}}{A sparse matrix with n rows and s columns, with 1 if the unit is in that stratification}

\item{\code{StrataFactor}}{Factor indicating strata}

\item{\code{StrataWeightRatio}}{For each unit, ratio of stratum weight to h_b; but see Details.}

\item{\code{Cluster}}{Factor indicating who's in the same cluster with who}

\item{\code{OriginalVariables}}{Look up table associating Covariates cols to terms in the calling formula, as in DesignMatrix}

\item{\code{Covariates}}{Numeric matrix, as in DesignMatrix, except here we presume columns to have been aligned (stratum-centered)}

\item{\code{NotMissing}}{matrix of element weights, normalized within stratum to have mean 1. (As otherwise, NAs are 0s)}
}}

