% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{alignedToInferentials}
\alias{alignedToInferentials}
\title{Align to Inferentials}
\usage{
alignedToInferentials(zz, tmat, ssn, ssvar, dv, wtsum)
}
\arguments{
\item{zz}{zz}

\item{tmat}{tmat}

\item{ssn}{ssn}

\item{ssvar}{ssvar}

\item{dv}{dv}

\item{wtsum}{wtsum}
}
\value{
list
}
\description{
Align to Inferentials
}
