% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{designToDescriptives}
\alias{designToDescriptives}
\title{Generate Descriptives}
\usage{
designToDescriptives(design, covariate.scaling = TRUE)
}
\arguments{
\item{design}{A Design}

\item{covariate.scaling}{Should covs be scaled?}
}
\value{
Descriptives
}
\description{
Use a design object to generate descriptive statistics that ignore
clustering. Stratum weights are respected.
}

