% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{designToDescriptives}
\alias{designToDescriptives}
\title{Generate Descriptives}
\usage{
designToDescriptives(design, covariate.scaling = NULL)
}
\arguments{
\item{design}{A DesignOptions object}

\item{covariate.scaling}{Scale estimates for covs, to use instead of internally calculated pooled SDs}
}
\value{
Descriptives
}
\description{
Use a design object to generate descriptive statistics that ignore clustering. 
Stratum weights are respected.
}

