% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{HB08}
\alias{HB08}
\title{Adjusted & combined differences as in Hansen & Bowers (2008)}
\usage{
HB08(alignedcovs)
}
\arguments{
\item{alignedcovs}{A CovsAlignedToADesign object}
}
\value{
list with components:
\describe{
  \item{z}{First item}
  \item{p}{Second item}
  \item{Msq}{Squared Mahalanobis distance of combined differences from origin}
  \item{DF}{degrees of freedom}
  \item{adj.mean.diffs}{Vector of sum statistics z'x-tilde, where x-tilde is the unit- and stratum-weighted covariate, with stratum centering.  This differs from the adjusted difference vector of Hansen & Bowers (2008) by a constant of proportionality.}
  \item{tcov}{Matrix of null covariances of Z'x-tilde vector, as above.}
}
}
\description{
Adjusted & combined differences as in Hansen & Bowers (2008)
}
\references{
Hansen, B.B. and Bowers, J. (2008), ``Covariate
  Balance in Simple, Stratified and Clustered Comparative
  Studies,'' \emph{Statistical Science} \bold{23}.
}
\seealso{
\code{\link{balanceTest}}, \code{\link{alignDesignsByStrata}}
}
\keyword{internal}
