% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{iqrDiff}
\alias{iqrDiff}
\title{IQR Difference}
\usage{
iqrDiff(q1 = 0.25, q2 = 0.75, type = 7)
}
\arguments{
\item{q1}{Lower quartile}

\item{q2}{Upper quartile}

\item{type}{type}
}
\value{
IQR
}
\description{
IQR Difference
}

