% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naImpute.R
\name{naImpute}
\alias{naImpute}
\title{Impute NA's}
\usage{
naImpute(FMLA, DATA, impfn = median, na.rm = TRUE,
  include.NA.flags = TRUE)
}
\arguments{
\item{FMLA}{Formula}

\item{DATA}{Data}

\item{impfn}{Function for imputing.}

\item{na.rm}{What to do with NA's}

\item{include.NA.flags}{Should NA flags be included}
}
\value{
Structure
}
\description{
Does this by doing something...
}
