% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotggbal.R
\name{plotggxbal}
\alias{plotggxbal}
\title{ggplot2 based balance plot}
\usage{
plotggxbal(x, xlab = "Standardized Differences", statistic = "std.diff",
  colors = NULL, symbols = NULL, absolute = FALSE, strata.labels = NULL,
  variable.labels = NULL, legend.position = "bottom", legend.title = "",
  var.order = NULL, ...)
}
\arguments{
\item{x}{Balance object}

\item{xlab}{Same as plot.xbal}

\item{statistic}{Same as plot.xbal}

\item{colors}{Vector of colors, same length as number of strata}

\item{symbols}{Vector of (numeric) symbols, same length as number of strata}

\item{absolute}{Same as plot.xbal}

\item{strata.labels}{Same as plot.xbal}

\item{variable.labels}{Same as plot.xbal}

\item{legend.position}{Where the legend should go ("bottom", "right")}

\item{legend.title}{Title of legend (blank by default)}

\item{var.order}{List of variable names (variable.labels if non-NULL) defining ordering. Otherwise alphabetical.}

\item{...}{}
}
\description{
ggplot2 based balance plot
}
