% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{quantileAbsoluteDifference}
\alias{quantileAbsoluteDifference}
\title{Quantile Absolute Difference}
\usage{
quantileAbsoluteDifference(quantiles)
}
\arguments{
\item{quantiles}{Quantile at which to calculate difference}
}
\value{
Function to calculate absolute value of difference between quantiles
}
\description{
Evaluates the difference at discrete points of the eCDF between treatment
and control group. Must specify quantile
}
\examples{
median.abs.diff <- quantileAbsoluteDifference(.50)
}
\seealso{
\code{\link{RItest}}
}
