% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{quantileDifference}
\alias{quantileDifference}
\title{Quantile Difference}
\usage{
quantileDifference(q = 0.5)
}
\arguments{
\item{q}{Quantile at which to calculate difference}
}
\value{
Function to calculate quantile difference
}
\description{
Evalutes difference between specific quantile between treatment and control
group. Uses median by default
}
\examples{
default <- quantileDifference() # median
custom <- quantileDifference(q = .65) # 65th percentile
}
\seealso{
\code{\link{RItest}}
}
