% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzpvalues.R
\name{Randomizationdistributionconfint}
\alias{Randomizationdistributionconfint}
\alias{confint.RandomizationDistribution}
\alias{plot.RandomizationDistributionConfInt}
\alias{summary.RandomizationDistributionConfInt}
\title{functions for RandomizationDistribution}
\usage{
\method{confint}{RandomizationDistribution}(object, parm, level = 0.95, ...)

summary.RandomizationDistributionConfInt(object, ...)

plot.RandomizationDistributionConfInt(object, ...)
}
\arguments{
\item{object}{RandomizationDistribution}

\item{parm}{param}

\item{level}{level}

\item{...}{Add'l arguments. Ignored.}
}
\description{
functions for RandomizationDistribution
}

