% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{rank.sum.functions}
\alias{rank.sum.functions}
\alias{rank.sum}
\alias{paired.sgnrank.sum}
\title{Rank based functions}
\usage{
rank.sum(ys, z)

paired.sgnrank.sum(ys, z)
}
\arguments{
\item{ys}{Outcome variable}

\item{z}{Treatment indicator vector that can be made logical}
}
\value{
Function to calculate rank-based sum
}
\description{
Various functions to compute rank based test statistics
}
\seealso{
\code{\link{RItest}}
}
