% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomization.distribution.R
\name{RItest}
\alias{RItest}
\title{Test a sharp hypothesis}
\usage{
RItest(y, z, test.stat, moe = NULL, parameters = NULL,
  sampler = simpleRandomSampler(z = z, b = rep(1, length(z))),
  samples = 5000, type = "exact", ...)
}
\arguments{
\item{y}{y}

\item{z}{z}

\item{test.stat}{test.stat}

\item{moe}{single function with signature f(y, z, param1, param2, etc.)odel of effect}

\item{parameters}{list of name = values, name = values, ...}

\item{sampler}{sampler}

\item{samples}{num samples}

\item{type}{type}

\item{...}{Add'l arguments}
}
\value{
SharpNullTest
}
\description{
Test a sharp hypothesis
}

