% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{scale.DesignOptions}
\alias{scale.DesignOptions}
\title{Scale DesignOptions}
\usage{
\method{scale}{DesignOptions}(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{DesignOptions object}

\item{center}{logical, or a function acceptable as \code{post.alignment.transform} arg of \code{alignDesignsByStrata()}}

\item{scale}{logical, whether to scale}
}
\description{
Scale DesignOptions
}
