% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{simpleRandomSampler}
\alias{simpleRandomSampler}
\title{Simple Random Sampler}
\usage{
simpleRandomSampler(total, treated, z, b)
}
\arguments{
\item{total}{size of the experimental pool}

\item{treated}{Total number to be assigned treatment. One of \code{treated} or \code{z} is required.}

\item{z}{Right now, z is binary or logical.}

\item{b}{records the membership in blocks within which treatment is assigned}
}
\value{
List
}
\description{
The simple random sampler creates a function that draws from the set of all randomizations with a fixed
number of treated units within blocks
}

