% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{slm_fit_csr}
\alias{slm_fit_csr}
\title{new version of slm.fit.csr}
\usage{
slm_fit_csr(x, y, ...)
}
\arguments{
\item{x}{As slm.fit.csr}

\item{y}{As slm.fit.csr}

\item{...}{As slm.fit.csr}
}
\value{
As slm.fit.csr
}
\description{
SparseM's slm.fit.csr has a bug for intercept only models
(admittedly, these are generally a little silly to be done as a
sparse matrix), but in order to avoid duplicate code, if
everything is in a single strata, we use the intercept only model. 
This implementation contains some workarounds to ensure that only
positive definite matrices are handed off to SparseM::chol()
}
