% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparseM_solve}
\alias{sparseM_solve}
\title{Helper function to slm.fit.csr.fixed}
\usage{
sparseM_solve(x, y, ...)
}
\arguments{
\item{x}{A slm.fit.csr}

\item{y}{A slm.fit.csr}

\item{...}{A slm.fit.csr}
}
\value{
list containing coefficients (vector or matrix) and Cholesky decomposition (of class matrix.csr.chol)
}
\description{
This function performs some checks and takes action to 
ensure positive definiteness of matrices passed to SparseM functions.
}
