% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SparseM_solve}
\alias{SparseM_solve}
\title{Helper function to slm_fit_csr}
\usage{
SparseM_solve(x, y, ...)
}
\arguments{
\item{x}{A slm.fit.csr}

\item{y}{A slm.fit.csr}

\item{...}{A slm.fit.csr}
}
\value{
list containing coefficients (vector or matrix), the Cholesky decomposition (of class matrix.csr.chol), and a vector specifying the indicies of which values on the diagonal of x'x are nonzero. These are named "coef", "chol" and "gramian_reduction_index", respectively.
}
\description{
This function performs some checks and takes action to 
ensure positive definiteness of matrices passed to SparseM functions.
}
