% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teststatistics.R
\name{test.stat.means}
\alias{fastmean}
\alias{fastvar}
\alias{mean.diff.lsfit}
\alias{mean.diff.vect}
\alias{mean.difference}
\alias{t.mean.difference}
\alias{test.stat.means}
\title{Mean functions}
\usage{
fastmean(ys)

\method{mean}{difference}(ys, z)

\method{mean}{diff.lsfit}(ys, z)

\method{mean}{diff.vect}(ys, z)

\method{t}{mean.difference}(ys, z)

fastvar(ys)
}
\arguments{
\item{ys}{ys}

\item{z}{Vector that can be made logical.}
}
\value{
mean
}
\description{
Various functions to compute means
}

