% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Design.R
\name{weightedDesign}
\alias{weightedDesign}
\title{Create a weighted Design}
\usage{
weightedDesign(design, stratum.weights = harmonic)
}
\arguments{
\item{design}{Design}

\item{stratum.weights}{Stratum weights
##' @return WeightedDesign}
}
\description{
Adds weights to existing Design.
}
