% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{XtX_pseudoinv_sqrt}
\alias{XtX_pseudoinv_sqrt}
\title{Matrix square root of XtX's pseudoinverse}
\usage{
XtX_pseudoinv_sqrt(mat, mat.is.XtX = FALSE, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{mat}{double-precision matrix}

\item{mat.is.XtX}{is mat a crossproduct of an X matrix, or X itself?}

\item{tol}{tolerance}
}
\value{
matrix of \code{ncol(mat)} rows and col rank (mat) columns
}
\description{
Modeled on MASS's \code{ginv}
}
\author{
Ben Hansen
}
\keyword{internal}
